[inkex] Rewrite inkex.errormsg tests to check output.
The old code, when referring to "Latin 1", seemed to in fact be talking
about literal (un-escpaed) Unicode characters. Passing actual Latin 1
characters to errormsg() causes them to be interpreted as UTF-8 and
produces garbage output.

Also, now that the output is checked programmatically, there's no need
to increase the verbosity of the test; use unittest.main() like most
other tests.