This is a personal project I have made which allows for button parameters to be made in .xml extension files.
Whenever the created button is clicked it sends a value of true and executes the current effect, this allows for the creation of an option in the Python file of an "inkbool" type with the name of the button to be monitoring the click state of the button and make different stuff depending if a specific button has been clicked or not.

In order to do this 2 new files must be created and changes in existing files must be made. I will send the current files I have which make this work.

These are the 2 files that create the button.