Add a font rendering and multilingue test based on Pepper&Carrot sources
This commit add the main file test-peppercarrot-text.svg: a collection of
speech-bubbles sampled from various language of the comic project Pepper&Carrot
(content: CC-By www.peppercarrot.com). To test render the file, a lot of fonts
were imported too: their license and copyright information (all compatible with
a Free Software project like Inkscape) were detailed in the License file, and
for longuer copyright infos about author and history of fork in a *.COPYRIGHT
file named in a similar way to the font project.

Renderings with my actual Inkscape were not provided: I'm affraid my environment
is polluted with other version of the same fonts installed, so it's better if
the base rendering in expected_rendering can be generated by someone else based
on the distribution of Fonts I shared here.

 + Add more ignore extension to the buildtool: *.ttc

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/1349