Moving the document replacement logic up a level
Needed that so that the Wasmer module can use it, and so put it
in the base implementation class. Probably long term should move
to somewhere in the document logic itself. But this is a smaller
transition.

Also, ran the code formatter which caused a bigger diff than this
would have been otherwise.