LIVE CODING ===================================================== LPE POWER STROKE linked to style dialog
Goal: A way to define a fill&stroke of a power stroke LPEs using standar fill and stroke dialog.

The main problem is powerstoke items are "visualy" a stroke but "realy" are a SVG filled path, so wen we apply styles to it we do the wrong applying inverted and no fill...

How to solve:
We have in diferent LPE a new feature to allow extra elements linked to main LPE item (meassure segments for example)

We take this feature and use to link fill based properties to LPE item and stroke based properties to the linked one.

The linked (stroke) item is unselectable so we have only the path of fill and the knots

STEPS DONE IN FIRST SESSION
1. Create fill & stroke items
2. Link style changed
3. Link also paint order

STEPS DONE IN MIDDLE
1. a very tyny refactor to well handle the LPE from master to a 1.0.x branch (currently master is unsuable to our goal)

STEPS TO DO IN NEXT MEETENG
1. Fix SPSHAPE BBox calculation on powerstroke items to include stroke in visual BBOX
2. Fix pen/pencil tool to work as current way using drop down toolbar shapes (triangle in...)
3. Use LPE versioning if necesary to make the code 1.0 and lower compatible