packaging/flatpak: add manifest for GNOME Builder integration
By adding the Flatpak manifest to Inkscape, GNOME Builder can automatically
discover it and use it to setup build containers. This allows potential
contributors to clone the Inkscape repository and click "Run".

At that point they have a C++ IDE with completion, diagnostics, and more
with nothing more than Builder and a clone of the Inkscape repository.