Resolve common caching with size for handles
Now the common rendered cache is built when asked with the size,
in this way the cache never really has anything extra and doesn't need
to be intialized according to any predetermined size. Infact, this shall
expectedly perform better than the previous (common cache)
implementation anyway.

Also, added the parsing map for other shapes and handles
previously not included.

As part of the project, "Customizable Appearance of Canvas Controls"