Document changes in Node Based Filter Editor Implemented
With this push, the Node Based Filter Editor is a lot more functional.
It can now create a filter and supports adding each primitive, and editing it's properties.
Connecting multiple primitives and changing their properties also reflects in the document.
Currently the following are yet to be implemented before it can replace the present editor
- Support for multiple filters (currently only first filter is supported, this shall be fixed by the next patch)
- Reading Filters from File
- Deletion, Duplication and Clipboard operations of nodes
- UI/UX is temporary and will be changed soon post discussion with UX team.