Improved UI, Added Deletion
The UI was improved temporarily to make it smaller, more readable
and use fonts according to theme and settings.
Additionally
- Added ability to delete selected nodes and primitives with right click
- Improved working of nodes with multiple inputs and document changes
- Improved document updating
- Working on duplication, and support for more filters.

The document should now reflect most of the changes in the node based
editor. The base for editor specific attributes such as updating the
positions of nodes on dragging, etc. has also been laid out.