Remove use of RegEx for splitting folders
Instead of blindly chunking a path using the directory symbol, we
want to split it into its actual components and we have an existing
function to do that.

Also fixed is the recent menu crashing if it's not found in the menu.ui.

And the prepending on .. to every recent file, no matter if it was in
the root directory or not. The code used to assume size 2 meant 1 folder
and 1 filename, but it's actually 1 root and 1 filename, so change the
checks to take advantage of that.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/5169