Fix regression in straight line on node deletion
The refactoring to the node tool's delete function caused
a regression where it was impossible to delete nodes and
have the resulting line be straight.

This adds in the missing functionality as well as fixing the
ctrl modifier setting which was the wrong way around.

This commit is compatible with 1.4.x and the next one will
fix the UX more.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/5183