Add a configurable undo stack size limit
Preference: A new integer preference added to Behaviour to specify the size of the
undo stack before discarding items. Default set to 200. A checkbox allows for the
current behaviour of infinate undos.

when commiting a new undo item we check for the size limit and remove items from
the back of the undo stack. This then calls a new notifyUndoExpired to all observers
which allows for the history dialog to remove expired items.

To improve the user interface, I change the text of the static item in the undo
dialog from [Unchanged] to [Changes forgotten] when we start reaching the limit
this then counts how many undo items have been forgotten using existing columns.

Also: Removes history_size from SPDocument which wasn't used for anything.