Don't reuse SPStyle when processing outline
Clearing the style caused the marker style to disapear. Making a new SPStyle
for the fill and the stroke fixes this issue.

Test case for stroke-to-path is enabled to catch this error in future.

Additionally, a bug in the marker handling multiple shape objects in the
SPMarker has been fixed. Using first_child is wrong according to the SVG
specification.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/5315