Add action to save the preferences
Allowing preferences to be saved while running can be helpful for
some situations for the user or the developer. This adds a button
in the preferences dialog next to "Reset Preferences".

As well as this we enable app actions to be used in extensions. The
ultimate goal is for extension authors to use this to ensure that
the live prefs are available to the extension author and can tell
something about the ongoing state of inkscape when it was called.