Fixes broken export to PDF with links.
When using a locale with an alternative decimal separator, Cairo 1.16 expects this decimal separator in the "rect" attribute that we use to create links in the PDF. This commit forces the use of the current locale's decimal separator with Cairo 1.16. Additionally it forces the use of the classic locale (and therefore of "." as a decimal separator) for later versions of Cairo.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/5354