Pre-multitab refactoring
All the boring and harmless changes split out from the tabbed ui commit.

The only questionable change is the removal in add_actions_view_mode(),
but this can be justified by saying there's no way to start up Inkscape
in fullscreen or focus mode, and in any case the settings are reapplied
later in view_set_gui().