Don't show icon in expander button for vertical toolbars
When a toolbar collapses, it displays an expander button consisting of
an icon with a dropdown arrow by its side.

This is problematic for vertical toolbars (only) because of the extra
width - GTK doesn't like it contradicting the previously reported width,
and it also makes the whole toolbar take up extra space.

The simplest solution is to keep the arrow, but hide the icon.