Fix Linux snap package build (1.4.x)
Make the snap build work again for 1.4.x.

Rework the patches for the file-open dialog and for launching extensions. Put these into the main code. Add a CMake option to enable/disable the snap-specific patches.

(Having a special compile flag for the snap package is quite ugly, however, it is only a temporary workaround for 1.4.
The upcoming code for 1.5 doesn't need the special flag anymore.)