Add tags to the sp-object modified flags
SP_OBJECT_MODIFIED_FLAG family is used extensively to signify that object
has been modified. There's no way to tell where the change is coming from though.
"User tags" can be added to the mix to mark senders of of those flags.
This is not as flexible as passing "this" pointer, but requires minimal modifications.
Existing code is already OR-ing flags as they are propagated, so passing tags is easy.
Dialogs listening for "selectionModified" can then distinguish senders as long as
they (senders) use unique tags.
To send tag, pass it into DocumentUndo::done or maybeDone calls.