Fix relative transformation actions
This is a regression that happened a long time ago when we moved to actions.

Previously inkscape allowed transforms to be anchored against whichever anchor
was being hovered over. Today, we have a click to select UX instead. But this
did not carry over to the transformations which are driven through g-actions.

Refactoring:

 * `setScaleRelative` is renamed `scaleRelative` to make it consistant with the
   other transformation relative functions.
 * Action specific functions moved from object-set, to the actions code itself
   removing a few uses of the desktop object in the object-set.
 * Functions added to get the desired anchor point from the selection easily.
 * Move screen zoom based actions to window actions so we don't have dangling
   app actions that don't work for the command line.
 * Remove rotate keys from the selection-tool and turn into actions like the
   scale operators.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/5456