Fix Export batch dialog crash / misbehavior with multiple documents open
Fixes the issue that switching between (nonempty) documents causes the export-batch dialog to show too more and more objects with every switch. The internal state became inconsistent in one edge case that was not triggered before.

As a lucky coincidence, it also fixes a crash when closing the export dialog (not sure why):

Fixes https://gitlab.com/inkscape/inbox/-/issues/11496