Fix a nasty hang on save causing all data loss
On specific versions of Gtk4 which upstream bugs related to the
filter selector. Inkscape would hang if the user browsed to a
folder that didn't contain any files which matched the filters
list.

This refactoring removes the use of object-based flags to record
state in a functional context. We use functions with explicit
names instead.