Fix recent files list in startup screen
The recent files list was improved in 1.4 to include the directory
when there are duplicate files. This is very useful but was missing
from the startup screen as the code for generating the list was not
generic.

This moves the recent files getter into it's own io files, this will
be useful for future development regarding recent files too.

Allows the map generator to run without needing to be called from
the menu builder.