Refactor the handling of sodipodi:nodetypes
The code of the Node Tool related to reading and writing
the XML attribute sodipodi:nodetypes is refactored and
enhanced with support for the new character 'e'. This
character triggers the creation of elliptical arc controls.
In other words, if the node type sequence, instead of,
"sc" is "sec", then the curve segment between the 's'-node
and the 'c'-node will have elliptical arc controls whenever
it is actually an arc.

The creation of nodes of different types is moved to a dedicated
factory class.