Create Elliptical Arc controllers only on demand
Ensure that elliptical arc controllers are created by the Node Tool
only when the character 'e' is found in the XML attribute sodipodi:
nodetypes or when the attribute is absent. Ensure that converting
an Ellipse to path manually sets this attribute to a string not
containing 'e's in order to preserve the historical behaviour of
the Node Tool on such paths.