Use latest lib2geom with GSL-based arc fit
Update lib2geom to the latest revision which includes functionality
for fitting an elliptical arc to the given tangent lines.

We also set -DHAVE_GSL as needed by lib2geom headers.

Note that the elliptical arc functionality checked out here requires
lib2geom to be built against GSL. In general, lib2geom always links
against GSL as soon as GSL is found by its CMake script; this is
always the case when building lib2geom standalone because GSL is a
REQUIRED package.

Therefore, the release packages for lib2geom are always GSL-enabled.

When building lib2geom from the internal submodule, we also have GSL,
since it is a REQUIRED dependency of Inkscape itself.