Fix UX issue with anchored rotate actions
Because multiple rotations change the bounding box multiple rotation
actions would produce an ever changing anchor point, we want to
remember the original anchor in the cases where we're using maybeDone
and therefore compounding the rotations together will use the same anchor.

Adds a function to SPDocument to get the action key for reading.