Refactor Splash Screen
There's a number of issues which are happening with the new splash screen.

 * Window isn't added to GApplication, causing macOS crashes
 * Splash isn't realised in time so don't work most of the time
 * Various conflicts with the start up screen itself

The transparencies are removed, the startup initialisation is repaired and
cleaned and the window is added to the GApp. I've also added a loading
label which fixes an empty area problem and provides a way to add messages
in the future if we want to.