Deprecate isValid() in preferences.cpp
isValid() was misleading as all it checked was if the value existed.
It has been deprecated in favour of of a renamed function isSet().

Move all instances of isValid() to use isSet() instead (preferences.cpp)

isValid*() functions are not used as the cost of using them is
same as the cost incurred by the get*() functions, and the get*
functions check the validity sunsequently anyway.