Add dedicated parsing utils for Selectors & CSS
The Selectors & CSS dialog currently attempts to parse a CSS stylesheet
naïvely, which is a source of many bugs.  This commit provides a set of
dedicated utilities which convert a raw CSS string into a rudimentary
"syntacic decomposition" consisting of individual elements of the CSS
syntax.

Currently, three types of elements of CSS syntax are handled:
* rule set statements, e.g., " circle, #sel { fill: red; } "
* @-rule statements, e.g., " @media print { /* any CSS content */ } "
* "other"/generic statements, e.g., @charset or @import.

The utilities provided here are tested with a new unit test.