Zoom with Middle Click Drag
Adds Zoom with Middle Click Drag as a feature. It makes use of the
Modifier Configuration and is disabled by default so as to not break
existing user workflows.

This also makes the rotate shortcut use the Modifier Configuration
Allowing a user to change the keybinds as per their preference.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/5080