Improve CLI tests with modular fixtures
 * Setup test creates the output or expected files
 * Main test checks the output, and only deletes what it creates
 * Clean test deletes the output and expected files

Using fixtures we can guarantee that the setup and cleanup will be
called when we ask for the main check test when they fail. The more
precise control over the deleting of files allows for more consistent
results when tests fail and reduce forgotten files.

The annoying 0.92 backup file is now deleted correctly fixing a long
standing issue.

The compare scripts are all improved so they can output the PNG file
with a known filename and have standard methods in utils for cleanup
and for keeping files when tests fail.

The CMake functions for building a test are split out for the pdfoutput
tests which are not included here. Also the inclusion of the get_pages
script which helps get a number of pages out of an SVG file without
calling Inkscape and slowing down make time.