Fix opening recent PDFs from Command Palette
When the user opens a recently opened PDF file from the Command Palette
and presses OK in the PDF import dialog, the PDF will now actually open.
Moreover, if the user presses Cancel in the PDF import dialog, Inkscape
will no longer crash.

Both buggy behaviours were triggered by the use of the wrong action
"app.file-open" meant for headless file opening, and are fixed by using
"app.file-open-window" to open the file in a window.

This is appropriate since the existence of an instance of a Command
Palette implies that we are in GUI mode.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/4413