Fix: ALT key hold events detection
The detection of ALT key presses was broken due to a typo error in the
definition of `INK_GDK_MODIFIER_MASK`. The `GDK_CONTROL_MASK` was
incorrectly duplicated, preventing the `GDK_ALT_MASK` from being
included in the modifier mask. This resulted in failing to recognize
ALT key hold events.

This commit corrects the definition of `INK_GDK_MODIFIER_MASK` to
include both `GDK_CONTROL_MASK` and `GDK_ALT_MASK`, resolving the issue.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/5555