Enable snapping to all baselines in multi-line text
This commit introduces functionality to allow snapping to all baselines
of a multi-line text object when the "Text baselines and anchors" snap
setting is enabled.  Previously, snapping was limited to the first
baseline in multi-line text.

This enhancement improves the usability of text snapping, allowing for
more accurate alignment of objects relative to multi-line text elements.

Fixes: https://gitlab.com/inkscape/inkscape/-/issues/5514