Fix crash on editing SVGs with text nodes
when 'Check Attributes on Editing' is enabled; The code requires the
node to be an element node. If it isn't, it tries to return an invalid
NULL ustring (its original return type was a c_str). We avoid this by
adding a node type check before calling it.

Moreover, the code now returns an empty string if it fails.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/5337