Snap package: Fix file permissions
Extensions, drag-and-drop and some other exotic functionalities do not work through xdg-portal.
As a workaround, we give the snap package direct access to the home directory and to removable media.

For exotic file paths like /tmp or /usr/share/doc this won't work but is also very difficult to fix.

Note that the connection to removable-media is not auto-activated by default; I will request this from the snap store admins once this is merged in.
Until then you need to run `snap connect inkscape:removable-media`

Partly fixes the following issues:
- https://gitlab.com/inkscape/inkscape/-/issues/3624
- https://gitlab.com/inkscape/inkscape/-/issues/3649
- https://gitlab.com/inkscape/inkscape/-/issues/2490