Allow multiline strings to expand in extension dialog
Effect-type extensions can add parameters to the effect dialog.
The string parameter supports the multiline property.

Despite setting this property, the entry would only occupy a single
line, even when more space is available. We fix this by adding
`set_propagate_natural_height` which allows the ScrolledWindow
to account for its children when it decides to expand or not.