Split off SPDesktop's dependent types into headers
Split off the class DesktopAffine and the enum CanvasFlip into separate
header files. This reduces the length of desktop.h from 579 to 485 lines
- more than 16%!

These entities are moved to the Inkscape namespace. The enum CanvasFlip
is converted to enum class. A bool member is reordered to reduce padding.

The includes in the header desktop.h are sorted and cleaned up. The only
other place using the enum Inkscape::CanvasFlip: the file actions-canvas
-transform.cpp, is updated and also has its includes cleaned up.

This refactoring will make it possible to one day write unit tests for
the class DesktopAffine.