Decouple unit test name from file name
Create an argument TEST_SOURCE for the CMake function
add_unit_test() in order to allow the developer to set
an arbitrary test main file. This will allow for the use
of subdirectories in the directory testfiles/src, which
in turn will allow for a better organization of tests.

Improve the documentation of the CMake setup for unit
tests and move it to where the developers are most likely
to see it.