Add a method to check equivalence of SPObjects
For the purposes of reusing SPObjects instead of creating new
ones which would be identical for all practical purposes (for
example to avoid duplicate swatches), it makes sense to introduce
a concept of "equivalence" of SPObjects.

We add a method isEquivalent() which can tell us whether this
SPObject is effectively equivalent to another one.