Rewrite SPDocument::importDefs()
The function for importing the <defs> from an external document into the
current one is rewritten. The rewrite takes advantage of the functions
defined in id-clash.cpp, some of which were not available at the time
when SPDocument code was written.

In addition to eliminating a separate, duplicate ID clash resolution
mechanism (which is useless when we already use id-clash.cpp), the logic
for special handling of certain SPObject types is eliminated.

This logic didn't really belong in the SPDocument; in fact, a document
is a "container" of sorts for many different objects and should not know
or care about the specific needs of these objects.

One way in which the logic of this function has been radically
simplified is that the resolution of ID clashes is no longer performed.
Instead, the function expects as a precondition that there are no shared
IDs between the current document and the source document from which to
import <defs>.