Decouple SPDocument from XML::SimpleDocument
Since the class SimpleDocument is a concrete implementation
class of the interface Inkscape::XML::Document, it should not
be visible to users of the interface, such as SPDocument.
This fixes a 4-year old TODO comment.

By decoupling SPDocument from the concrete implementation, we
are also improving testability of SPDocument.

In addition, a method _duplicate() added as a stop-gap solution
in SimpleNode is eliminated. In a well-designed inheritance
hierarchy, a class should not serve as both a base class and a
concrete implementation class.