Refactor version reading in SPRoot
Reading and writing version strings needs to be something that is internal
to SPRoot, not handled by an external file. Likewise writing versions
should be handled by the new pre-processing actions and SPRoot.

Along with this we reduce the complexity of the version system removing
the unused "original" concept and simply parsing and saving the versions.