Refactor color-interpolation so we can use color spaces
We want to do more color interpolation than just the SVG standard
values, so we want to refactor the enums currently used into being
the Color Space system directly.

We add a interpolation attribute value matching and interpolation check
into the color space API itself and a manager function to parse it out.

We then move SPStyle to use the new parser instead of being an enum.

Interation into the rendering code will come later, so we keep the old
enum for just those parts until the time it can be replaced.