Prevent export-batch-path attribute from modifying document on load
If the user has the export dialog open, it will pollute any opened
document with the batch path attribute without any undo. This does
not affect the CLI but it does mean a user could leak directory
information to new documents or other documents loaded.

We ONLY ever want to save this information into the document on
a successful export operation.