Adjust cursor shadows so they never get cut off
Any shared filter that has a blur filter effect should not use
filterUnits based on the ObjectBoundingBox since the blur's spread
is related to it's stdDeviation and not the object's size.

To fix this we set the filterUnits to userSpaceOnUse and the
coordinates to 0,0 32,32 which covers the entire cursor canvas.

This fixes any future issues with shadows in the Dash cursor set.

Fixes https://gitlab.com/inkscape/inkscape/-/issues/5849