Adjust environment setup to new bundle structure
We have replaced the barely maintained gtk-mac-bundler with a new tool,
abcreate. The new bundler follows Apple's guidelines more closely in
regards to "what goes where", bringing changes to the bundle layout.

This was supposed to be part of 91c004f230cef15de195c0b36b953d11ebdd0415
but was forgotten:
  - Prefix to the bundle is Contents, not Contents/Resources. Everything
    in a bundle lives below Contents.
  - Libraries live in Contents/Frameworks.
  - Binaries live in Contents/MacOS, Contents/Resources/bin is being
    phased out.