Fix the incorrect typing of CMS spaces
Each color space has two types, one is it's actual type such as
RGB, CMS, CSSNAME etc. This is synonymous with the class type and
in future we may dissolve this into type checking. C++ permitting.

The second is the component type, which is what channels are actually
in the color space. Mostly this is the same as the first type. But
for CSSNAME and CMS, it is not. The CMS type was incorrectly claiming
to be RGB which meant in practice any RGB lookup could have returned
a CMS space instead of the expected sRGB default profile.

 * Added an equality function to reduce API verbosity.
 * Also removed a debug print.