Fix PDF input color support
We want to be able to support more than just RGB and icc profiles in
pdf input and in order to do this we have to refactor the profile
creation, name generation as well as propperly handle the intent.

Our output now correctly applies document icc profiles as well as
device-cmyk with and without icc profiles.

 * Color API has been improved with better profile name handling
 * Add Profile::create_from_copy to ensure lcms2 memory from poppler
   is respected.
 * Add an option to convert to sRGB which is default to true.
 * Add convert colors to the CLI, the default is false.