Fix UB due to incomplete type deletion
Several classes have an implicit or defaulted destructor but also
contain a unique_ptr to a forward-declared type. E.g. InkviewWindow
containing a pointer to SPDocument.

Somehow none of the CI builders complain about this, but it does break
building with very recent clang, and it is absolutely UB.

Fix it by moving those destructors to the cpp file.