Reduce text pick-box size to allow easier picking of background.
This changes the pick box from using the entire font design space (em-box height) to using
the "inked" box, subject to a minimum size of the glyph advance in width and half the em-box
in height.

Also, increase pick-box by one screen pixel to make it easier to pick text when zoomed out.

Fixes #5846.