Import preview page when using Cairo PDF import.
Cairo import can only import one page at a time. It is not clear
how to select which page is imported in the PDF Import dialog. The
most obvious way to select the page is with the preview controls.

Note: This does not make sense to use this method for internal
PDF import as more than one page can be imported at a time.