CMake Unit test TEST_SOURCE argument ignored during construction
The TEST_SOURCE argument must be prefixed with ARG_ as it is a parsed
function argument. That this wasn't caught seems to be because none
of the unit tests actually use this argument. To avoid duplicate
sources being added we else conditional else and add some warnings
to provide visability to future test writers.