Use a self-hosted Windows runner for 7z build
This change significantly reduces build times for the Windows 7z
artifact within the Inkscape group. Non-members will fall back to
GitLab's SaaS Windows fleet as they won't have access to this runner.

The exe and msi builds also remain on GitLab's fleet and will stay
there. They are not to be moved.

Move from 7z-packaged dependencies to a mountable wim image. This is
a huge benefit for a persistent runner and still a small one for an
ephemeral one (mounting is still faster than 7z extraction).

Fix a library name to unbreak packaging.

Increase self-termination timeout for jobs on GitLab's fleet to
100 minutes.