Fix CSS tweaking when theme contrast is not default (< 10)
sp_tweak_background_colors: background-image rules were rewritten with an extra image() in the value.
Removing the extra image() fixes the red background on affected UI elements (mainly buttons) when
contrast is set to < 10.

ThemeContext::add_gtk_css: Replace some nullptrs with empty strings to avoid assertion failures.

The above fixes https://gitlab.com/inkscape/inkscape/-/issues/5434

ThemeContext::add_gtk_css: When loading the CSS for a theme, the variant should be set by whether the
current theme is actually dark, not whether it's preferred dark. Fixes broken colours when
contrast < 10 and the actual darkness of the theme does not match the preferDarkTheme setting
(for example, when the theme only has a dark option regardless of what the user prefers).