Remove Application refcounting
Application was leaking due to the last ref never being dropped. Get rid
of the refcounting entirely, and make it live in a function-local static
variable instead, which is destroyed just after end of main.

Documents therefore no longer need to hold a ref on Application, because
Documents are guaranteed to be destroyed before end of main. Thus the
keepalive flag can be removed from SPDocument.

Note: All nontrivial changes are in inkscape.cpp and document.cpp; the
bulk of the commit is repetetive 'keepalive' flag removal.

Minor:

* Remove a comment saying this would break the Filters/Extensions menus;
  that issue seems to have been fixed.

* Downgrade "desktops still in list on application destruction!" from
  error to critical. It can happen on executing "quit immediately".