Rework when Preferences::save() is called
Rather than saving Preferences when the last desktop is closed, save
them after InkscapeApplication exits, which is slightly later but still
before the end of main.

Don't have Preferences::unload() call save() by default. This is because
save() is not safe to run after end of main (issues with statics).

Minor:

* Drop a useless Preferences::unload(false) in the crash handler.
  Preferences aren't saved on crash anyway.

* InkscapeApplication::exit() no longer does anything after this, and
  can be removed.

* We no longer save prefs on exiting Inkview. This should be fine as
  Inkview doesn't allow changing the prefs.

* preferences-test.cpp fallout is because that test relied on stored
  preferences state between test runs.