Add script to setup full Windows dev environment
Add a reproducible "one-click" alternative available to set up a whole development environment for Inkscape on Windows:

1. Set up VSCode
2. Set up Inkscape dependencies
3. Check out and compile Inkscape

We offer a script to run on any Windows PC, and also a VM configuration for Vagrant to set up a complete Win11 development VM from scratch.

**Motivation**:

Users have to follow several manual steps to get started with Inkscape development on Windows or to reproduce problems with the Windows build dependencies.
Before this change, only the dependency installation (step 2 from the above list) was automated, the rest was manual. Now, we offer a script that does everything.

Use-Cases:
1. As an experienced developer, I want to triage issues that the build instructions are not working: Just run a one-line command `vagrant destroy; vagrant up`. After two hours of waiting I have a fresh VM with the same result as if I had manually followed through all the instructions in `windows.md`.

2. As a new and unexperienced developer, I want to play around debugging Inkscape on Windows but I am overwhelmed by the long instructions. Now I can just click on `windows-installFullDevEnv-clickHere.bat`, do something else for two hours, and come back to a VSCode  window with Inkscape source and press F5 to start debugging.

3. I don't want to use the new magic stuff but still do it manually. This is still possible.